auth_proxy: Added AuthType to SessionState and forwarded headers.
Encodes the auth type(oauth2, auth_api, basic_auth) into the session
state and cookie. Also forwards this information to the upstream via the
X-Forwarded-Auth-Type header. Now that we keep track of the auth type,
we can avoid hitting the auth_api unnecessarily if the user
authenticated via oauth or basic auth.